requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
pillow>=11.0.0
numpy>=2.0.0
inference-sdk==0.9.11
openai>=1.0.0
